fleshed out the base classes to have method definitions
THIS BASE CODE IS SUPER ROUGH

It compiles but must be reviewed and rewritten extensively in
order to work properly. It is just a starting off point.

Furthermore, I had to make a lot of assumptions when filling
in variables and defining methods; if I am mistaken, please
feel free to refactor this code as is necessary in order to
meet the specs.